requirements.txt
flask
flask-cors
chromadb
sentence-transformers
groq
pymupdf
python-dotenv
werkzeug
gunicorn
